requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.30
koleo-cli>=0.2.137
aiohttp>=3.9
pydantic>=2.5
